Core: Split Action per-class, collapse Sha256 + Truncate subclasses
Mirrors the org.apache.iceberg.transforms.* layout: one file per concrete
action instead of 10 nested classes in Action.java. Action.java now holds
only the interface plus the BaseAction abstract (fieldId carrier).

Algorithmic simplifications along the way:
- ShowLast4 rewritten single-pass with a 4-offset ring buffer
  (was O(2n) from a separate codePointCount + mask loop).
- Sha256 4 per-type subclasses (Sha256String/Integer/Long/Binary)
  collapsed into one Sha256Fn with a Codec enum carrying the update/
  encode pair.
- Truncate 3 per-storage subclasses (Date/Timestamp/TimestampNano)
  collapsed into one TruncateTemporalFn with Unit + Storage enums.
- mapCodePoint moved from Actions to MaskAlphanum.maskCodePoint;
  matches the spec phrasing "redacts the remainder using mask-alphanum
  rules" that ShowFirst4/ShowLast4 reference.

UnknownAction.bind() now throws IllegalArgumentException for consistency
with the other actions' bind-time type rejection.

35 TestActions tests pass, spotless + revapi clean.
