+++ core/src/main/java/org/apache/iceberg/functions/ApplyExpression.java
@@ -1,0 +1,69 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.iceberg.functions;
+
+import org.apache.iceberg.expressions.Expression;
+import org.apache.iceberg.relocated.com.google.common.base.Preconditions;
+import org.apache.iceberg.types.Type;
+import org.apache.iceberg.util.SerializableFunction;
+
+/**
+ * Replaces the column value with the result of a server-provided Expression. Not supported by this
+ * client yet — Iceberg Expressions are currently boolean-only, so binding returns a function that
+ * throws on apply.
+ */
+public final class ApplyExpression extends Action.BaseAction<Object, Object> {
+  private final Expression expression;
+
+  public ApplyExpression(int fieldId, Expression expression) {
+    super(fieldId);
+    Preconditions.checkArgument(expression != null, "Invalid expression: null");
+    this.expression = expression;
+  }
+
+  public Expression expression() {
+    return expression;
+  }
+
+  @Override
+  public String actionType() {
+    return APPLY_EXPRESSION;
+  }
+
+  @Override
+  public boolean canBind(Type type) {
+    return true;
+  }
+
+  @Override
+  public SerializableFunction<Object, Object> bind(Type type) {
+    return ApplyExpressionFn.INSTANCE;
+  }
+
+  private static final class ApplyExpressionFn implements SerializableFunction<Object, Object> {
+    static final ApplyExpressionFn INSTANCE = new ApplyExpressionFn();
+
+    @Override
+    public Object apply(Object value) {
+      throw new UnsupportedOperationException(
+          "apply-expression column projection is not supported by this client "
+              + "(Iceberg Expression is currently boolean-only)");
+    }
+  }
+}
