+++ core/src/test/java/org/apache/iceberg/functions/TestActions.java
@@ -37,42 +37,38 @@ public class TestActions {
 
   @Test
   public void maskAlphanumSpecExample() {
-    SerializableFunction<String, String> fn =
-        new Action.MaskAlphanum(1).bind(Types.StringType.get());
+    SerializableFunction<String, String> fn = new MaskAlphanum(1).bind(Types.StringType.get());
     assertThat(fn.apply("prashant010696@gmail.com")).isEqualTo("xxxxxxxxnnnnnn@xxxxx.xxx");
   }
 
   @Test
   public void maskAlphanumPreservedPunctuation() {
-    SerializableFunction<String, String> fn =
-        new Action.MaskAlphanum(1).bind(Types.StringType.get());
+    SerializableFunction<String, String> fn = new MaskAlphanum(1).bind(Types.StringType.get());
     assertThat(fn.apply("(555) 123-4567")).isEqualTo("(nnn)xnnn-nnnn");
     assertThat(fn.apply("a.b,c")).isEqualTo("x.x,x");
   }
 
   @Test
   public void maskAlphanumNullInNullOut() {
-    SerializableFunction<String, String> fn =
-        new Action.MaskAlphanum(1).bind(Types.StringType.get());
+    SerializableFunction<String, String> fn = new MaskAlphanum(1).bind(Types.StringType.get());
     assertThat(fn.apply(null)).isNull();
   }
 
   @Test
   public void maskAlphanumEmptyString() {
-    SerializableFunction<String, String> fn =
-        new Action.MaskAlphanum(1).bind(Types.StringType.get());
+    SerializableFunction<String, String> fn = new MaskAlphanum(1).bind(Types.StringType.get());
     assertThat(fn.apply("")).isEqualTo("");
   }
 
   @Test
   public void showFirst4SpecExample() {
-    SerializableFunction<String, String> fn = new Action.ShowFirst4(1).bind(Types.StringType.get());
+    SerializableFunction<String, String> fn = new ShowFirst4(1).bind(Types.StringType.get());
     assertThat(fn.apply("prashant010696@gmail.com")).isEqualTo("prasxxxxnnnnnn@xxxxx.xxx");
   }
 
   @Test
   public void showFirst4FourOrFewerReturnedUnchanged() {
-    SerializableFunction<String, String> fn = new Action.ShowFirst4(1).bind(Types.StringType.get());
+    SerializableFunction<String, String> fn = new ShowFirst4(1).bind(Types.StringType.get());
     assertThat(fn.apply("abcd")).isEqualTo("abcd");
     assertThat(fn.apply("ab")).isEqualTo("ab");
     assertThat(fn.apply("")).isEqualTo("");
@@ -80,61 +76,55 @@ public class TestActions {
 
   @Test
   public void showLast4SpecExample() {
-    SerializableFunction<String, String> fn = new Action.ShowLast4(1).bind(Types.StringType.get());
+    SerializableFunction<String, String> fn = new ShowLast4(1).bind(Types.StringType.get());
     assertThat(fn.apply("4111-1111-1111-4444")).isEqualTo("nnnn-nnnn-nnnn-4444");
   }
 
   @Test
   public void showLast4FourOrFewerReturnedUnchanged() {
-    SerializableFunction<String, String> fn = new Action.ShowLast4(1).bind(Types.StringType.get());
+    SerializableFunction<String, String> fn = new ShowLast4(1).bind(Types.StringType.get());
     assertThat(fn.apply("abcd")).isEqualTo("abcd");
     assertThat(fn.apply("ab")).isEqualTo("ab");
   }
 
   @Test
   public void replaceWithNullAlwaysReturnsNull() {
-    SerializableFunction<Object, Object> fn =
-        new Action.ReplaceWithNull(1).bind(Types.IntegerType.get());
+    SerializableFunction<Object, Object> fn = new ReplaceWithNull(1).bind(Types.IntegerType.get());
     assertThat(fn.apply(42)).isNull();
     assertThat(fn.apply(null)).isNull();
 
     SerializableFunction<Object, Object> strFn =
-        new Action.ReplaceWithNull(1).bind(Types.StringType.get());
+        new ReplaceWithNull(1).bind(Types.StringType.get());
     assertThat(strFn.apply("hello")).isNull();
   }
 
   @Test
   public void maskToFixedValueString() {
-    SerializableFunction<Object, Object> fn =
-        new Action.MaskToFixedValue(1).bind(Types.StringType.get());
+    SerializableFunction<Object, Object> fn = new MaskToFixedValue(1).bind(Types.StringType.get());
     assertThat(fn.apply("anything")).isEqualTo("XXXXXXXX");
   }
 
   @Test
   public void maskToFixedValueInt() {
-    SerializableFunction<Object, Object> fn =
-        new Action.MaskToFixedValue(1).bind(Types.IntegerType.get());
+    SerializableFunction<Object, Object> fn = new MaskToFixedValue(1).bind(Types.IntegerType.get());
     assertThat(fn.apply(42)).isEqualTo(0);
   }
 
   @Test
   public void maskToFixedValueLong() {
-    SerializableFunction<Object, Object> fn =
-        new Action.MaskToFixedValue(1).bind(Types.LongType.get());
+    SerializableFunction<Object, Object> fn = new MaskToFixedValue(1).bind(Types.LongType.get());
     assertThat(fn.apply(42L)).isEqualTo(0L);
   }
 
   @Test
   public void maskToFixedValueDouble() {
-    SerializableFunction<Object, Object> fn =
-        new Action.MaskToFixedValue(1).bind(Types.DoubleType.get());
+    SerializableFunction<Object, Object> fn = new MaskToFixedValue(1).bind(Types.DoubleType.get());
     assertThat(fn.apply(3.14)).isEqualTo(0.0d);
   }
 
   @Test
   public void maskToFixedValueBoolean() {
-    SerializableFunction<Object, Object> fn =
-        new Action.MaskToFixedValue(1).bind(Types.BooleanType.get());
+    SerializableFunction<Object, Object> fn = new MaskToFixedValue(1).bind(Types.BooleanType.get());
     assertThat(fn.apply(true)).isEqualTo(false);
   }
 
@@ -142,8 +132,7 @@ public class TestActions {
   public void maskToFixedValueDate() {
     int input = DateTimeUtil.daysFromDate(LocalDate.of(2024, 7, 15));
     int expected = DateTimeUtil.daysFromDate(LocalDate.of(1970, 1, 1));
-    SerializableFunction<Object, Object> fn =
-        new Action.MaskToFixedValue(1).bind(Types.DateType.get());
+    SerializableFunction<Object, Object> fn = new MaskToFixedValue(1).bind(Types.DateType.get());
     assertThat(fn.apply(input)).isEqualTo(expected);
   }
 
@@ -153,14 +142,13 @@ public class TestActions {
         LocalDateTime.of(2024, 7, 15, 13, 45, 30).toEpochSecond(ZoneOffset.UTC) * 1_000_000L;
     long expected = LocalDateTime.of(1970, 1, 1, 0, 0).toEpochSecond(ZoneOffset.UTC) * 1_000_000L;
     SerializableFunction<Object, Object> fn =
-        new Action.MaskToFixedValue(1).bind(Types.TimestampType.withZone());
+        new MaskToFixedValue(1).bind(Types.TimestampType.withZone());
     assertThat(fn.apply(input)).isEqualTo(expected);
   }
 
   @Test
   public void maskToFixedValueBinary() {
-    SerializableFunction<Object, Object> fn =
-        new Action.MaskToFixedValue(1).bind(Types.BinaryType.get());
+    SerializableFunction<Object, Object> fn = new MaskToFixedValue(1).bind(Types.BinaryType.get());
     ByteBuffer result = (ByteBuffer) fn.apply(ByteBuffer.wrap(new byte[] {1, 2, 3}));
     assertThat(result.remaining()).isEqualTo(0);
   }
@@ -168,7 +156,7 @@ public class TestActions {
   @Test
   public void maskToFixedValueDecimal() {
     SerializableFunction<Object, Object> fn =
-        new Action.MaskToFixedValue(1).bind(Types.DecimalType.of(10, 2));
+        new MaskToFixedValue(1).bind(Types.DecimalType.of(10, 2));
     BigDecimal result = (BigDecimal) fn.apply(new BigDecimal("12.34"));
     assertThat(result.compareTo(BigDecimal.ZERO)).isEqualTo(0);
     assertThat(result.scale()).isEqualTo(2);
@@ -176,8 +164,7 @@ public class TestActions {
 
   @Test
   public void maskToFixedValueNullReturnsNull() {
-    SerializableFunction<Object, Object> fn =
-        new Action.MaskToFixedValue(1).bind(Types.IntegerType.get());
+    SerializableFunction<Object, Object> fn = new MaskToFixedValue(1).bind(Types.IntegerType.get());
     assertThat(fn.apply(null)).isNull();
   }
 
@@ -185,8 +172,7 @@ public class TestActions {
   public void truncateToYearDate() {
     int input = (int) LocalDate.of(2024, 7, 15).toEpochDay();
     int expected = (int) LocalDate.of(2024, 1, 1).toEpochDay();
-    SerializableFunction<Object, Object> fn =
-        new Action.TruncateToYear(1).bind(Types.DateType.get());
+    SerializableFunction<Object, Object> fn = new TruncateToYear(1).bind(Types.DateType.get());
     assertThat(fn.apply(input)).isEqualTo(expected);
   }
 
@@ -194,8 +180,7 @@ public class TestActions {
   public void truncateToMonthDate() {
     int input = (int) LocalDate.of(2024, 7, 15).toEpochDay();
     int expected = (int) LocalDate.of(2024, 7, 1).toEpochDay();
-    SerializableFunction<Object, Object> fn =
-        new Action.TruncateToMonth(1).bind(Types.DateType.get());
+    SerializableFunction<Object, Object> fn = new TruncateToMonth(1).bind(Types.DateType.get());
     assertThat(fn.apply(input)).isEqualTo(expected);
   }
 
@@ -206,7 +191,7 @@ public class TestActions {
     long expectedMicros =
         LocalDateTime.of(2024, 1, 1, 0, 0, 0).toEpochSecond(ZoneOffset.UTC) * 1_000_000L;
     SerializableFunction<Object, Object> fn =
-        new Action.TruncateToYear(1).bind(Types.TimestampType.withZone());
+        new TruncateToYear(1).bind(Types.TimestampType.withZone());
     assertThat(fn.apply(inputMicros)).isEqualTo(expectedMicros);
   }
 
@@ -217,14 +202,13 @@ public class TestActions {
     long expectedMicros =
         LocalDateTime.of(2024, 7, 1, 0, 0, 0).toEpochSecond(ZoneOffset.UTC) * 1_000_000L;
     SerializableFunction<Object, Object> fn =
-        new Action.TruncateToMonth(1).bind(Types.TimestampType.withZone());
+        new TruncateToMonth(1).bind(Types.TimestampType.withZone());
     assertThat(fn.apply(inputMicros)).isEqualTo(expectedMicros);
   }
 
   @Test
   public void sha256GlobalStringIsDeterministic() {
-    SerializableFunction<Object, Object> fn =
-        new Action.Sha256Global(1).bind(Types.StringType.get());
+    SerializableFunction<Object, Object> fn = new Sha256Global(1).bind(Types.StringType.get());
     String first = (String) fn.apply("hello");
     String second = (String) fn.apply("hello");
     assertThat(first).isEqualTo(second);
@@ -233,16 +217,14 @@ public class TestActions {
 
   @Test
   public void sha256GlobalBinaryReturns32Bytes() {
-    SerializableFunction<Object, Object> fn =
-        new Action.Sha256Global(1).bind(Types.BinaryType.get());
+    SerializableFunction<Object, Object> fn = new Sha256Global(1).bind(Types.BinaryType.get());
     ByteBuffer result = (ByteBuffer) fn.apply(ByteBuffer.wrap(new byte[] {1, 2, 3}));
     assertThat(result.remaining()).isEqualTo(32);
   }
 
   @Test
   public void sha256GlobalIntegerDeterministic() {
-    SerializableFunction<Object, Object> fn =
-        new Action.Sha256Global(1).bind(Types.IntegerType.get());
+    SerializableFunction<Object, Object> fn = new Sha256Global(1).bind(Types.IntegerType.get());
     Object first = fn.apply(42);
     Object second = fn.apply(42);
     assertThat(first).isEqualTo(second);
@@ -251,7 +233,7 @@ public class TestActions {
 
   @Test
   public void sha256GlobalLongDeterministic() {
-    SerializableFunction<Object, Object> fn = new Action.Sha256Global(1).bind(Types.LongType.get());
+    SerializableFunction<Object, Object> fn = new Sha256Global(1).bind(Types.LongType.get());
     Object first = fn.apply(42L);
     Object second = fn.apply(42L);
     assertThat(first).isEqualTo(second);
@@ -265,16 +247,15 @@ public class TestActions {
     Arrays.fill(saltA, (byte) 1);
     Arrays.fill(saltB, (byte) 2);
     SerializableFunction<Object, Object> fnA =
-        new Action.Sha256QueryLocal(1).bind(Types.StringType.get(), saltA);
+        new Sha256QueryLocal(1).bind(Types.StringType.get(), saltA);
     SerializableFunction<Object, Object> fnB =
-        new Action.Sha256QueryLocal(1).bind(Types.StringType.get(), saltB);
+        new Sha256QueryLocal(1).bind(Types.StringType.get(), saltB);
     assertThat(fnA.apply("hello")).isNotEqualTo(fnB.apply("hello"));
   }
 
   @Test
   public void sha256QueryLocalSaltMustBeAtLeast16Bytes() {
-    assertThatThrownBy(
-            () -> new Action.Sha256QueryLocal(1).bind(Types.StringType.get(), new byte[15]))
+    assertThatThrownBy(() -> new Sha256QueryLocal(1).bind(Types.StringType.get(), new byte[15]))
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessageContaining("16 bytes");
   }
@@ -282,7 +263,7 @@ public class TestActions {
   @Test
   public void applyExpressionFailsOnApply() {
     SerializableFunction<Object, Object> fn =
-        new Action.ApplyExpression(1, Expressions.alwaysTrue()).bind(Types.StringType.get());
+        new ApplyExpression(1, Expressions.alwaysTrue()).bind(Types.StringType.get());
     assertThatThrownBy(() -> fn.apply("any"))
         .isInstanceOf(UnsupportedOperationException.class)
         .hasMessageContaining("apply-expression");
@@ -290,36 +271,34 @@ public class TestActions {
 
   @Test
   public void bindRejectsMaskAlphanumOnNonString() {
-    assertThatThrownBy(() -> new Action.MaskAlphanum(1).bind(Types.IntegerType.get()))
+    assertThatThrownBy(() -> new MaskAlphanum(1).bind(Types.IntegerType.get()))
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessageContaining("STRING");
   }
 
   @Test
   public void bindRejectsTruncateOnUnsupportedType() {
-    assertThatThrownBy(() -> new Action.TruncateToYear(1).bind(Types.StringType.get()))
+    assertThatThrownBy(() -> new TruncateToYear(1).bind(Types.StringType.get()))
         .isInstanceOf(IllegalArgumentException.class)
         .hasMessageContaining("not supported for type");
   }
 
   @Test
   public void bindFailsClosedOnUnknownAction() {
-    assertThatThrownBy(() -> new Action.Unknown(1, "future-mask-v2").bind(Types.StringType.get()))
-        .isInstanceOf(IllegalStateException.class)
+    assertThatThrownBy(() -> new UnknownAction(1, "future-mask-v2").bind(Types.StringType.get()))
+        .isInstanceOf(IllegalArgumentException.class)
         .hasMessageContaining("future-mask-v2");
   }
 
   @Test
   public void sha256NullInNullOut() {
-    SerializableFunction<Object, Object> fn =
-        new Action.Sha256Global(1).bind(Types.StringType.get());
+    SerializableFunction<Object, Object> fn = new Sha256Global(1).bind(Types.StringType.get());
     assertThat(fn.apply(null)).isNull();
   }
 
   @Test
   public void truncateNullInNullOut() {
-    SerializableFunction<Object, Object> fn =
-        new Action.TruncateToYear(1).bind(Types.DateType.get());
+    SerializableFunction<Object, Object> fn = new TruncateToYear(1).bind(Types.DateType.get());
     assertThat(fn.apply(null)).isNull();
   }
 }
